TJFE-61 [refactor] kakaoDTO refactor
소셜 로그인 dto 통합으로 인하여 카카오 dto 수정
이름 가져오던 부분 삭제
가독성을 위해 패키지 생성
